requirements.txt
# ===== ANÁLISE DE DADOS =====
pandas
numpy

# ===== FONTES DE DADOS FINANCEIROS =====
finnhub-python
python-bcb
investpy
fredapi
yfinance
Nasdaq-Data-Link
alpha-vantage

# ===== INTERFACE DE TERMINAL =====
rich
blessed
tabulate

# ===== VISUALIZAÇÃO =====
plotext
matplotlib
plotly

# ===== INDICADORES TÉCNICOS =====
ta

# ===== MACHINE LEARNING =====
scikit-learn
tensorflow
keras
joblib

# ===== SÉRIES TEMPORAIS =====
prophet
statsmodels

# ===== BACKTESTING =====
backtesting

# ===== UTILITÁRIOS =====
requests
python-dotenv
plyer

# ===== DESENVOLVIMENTO E TESTES =====
pytest
pytest-cov
pyinstaller